- Fix path resolve - Added apply command
