+++ webpack.config.js
@@ -1,29 +1,29 @@
-var path = require('path');
+var path = require('path')
 
 module.exports = {
   entry: './src/index.js',
   output: {
     path: path.resolve(__dirname, 'lib'),
-		filename: 'bundle.js',
-		libraryTarget: 'commonjs2'
-	},
-	resolve: {
-		extensions: ['*', '.js', '.jsx']
-	},
-	module: {
-		rules: [
+    filename: 'bundle.js',
+    libraryTarget: 'commonjs2'
+  },
+  resolve: {
+    extensions: ['*', '.js', '.jsx']
+  },
+  module: {
+    rules: [
       {
-				test: /\.jsx$/,
-				loader: 'babel-loader',
+        test: /\.jsx$/,
+        loader: 'babel-loader',
         include: path.resolve(__dirname, 'src'),
-        exclude: /(node_modules|bower_components)/,
+        exclude: [/node_modules/, /bower_components/]
       }
     ]
-	},
-	node: {
-		fs: "empty"
- 	},
-	externals: {
-    'react': 'commonjs react'
-	}
-};
+  },
+  node: {
+    fs: 'empty'
+  },
+  externals: {
+    react: 'commonjs react'
+  }
+}
